feat(gtpu-dataplane): gate the OpenShift 4.18 kernel and expose a load-capability API (#544)
* feat(gtpu-dataplane): expose a committed-classifier load capability

Consumers had no supported way to ask whether the committed classifiers
load on a node, so a readiness gate had to compare uname against 6.8 or
read a bpftool helper inventory. The first refuses every enterprise
kernel that backports bpf_loop regardless of whether the object loads
there; the second reports helper names and cannot see a verifier
rejection, so it admits nodes whose only symptom is zero forwarding.

probe_committed_classifier_load attempts the real load and returns
Loadable, VerifierRejected or UnableToAttempt. It runs the same
EbpfLoader invocation and the same program load as the attach path, so
the answer is about the committed object rather than a synthetic probe,
and it stops before attaching: no clsact qdisc is created and no filter
is installed, which matters because the attach path never removes the
qdisc it adds. Map pins go to a per-attempt directory that is removed on
every path, so the probe is safe beside live traffic.

Verifier rejection and an unattemptable environment stay distinct
because they tell an operator different things: missing bpffs, missing
BTF or absent CAP_BPF are reported as blockers, never as a verdict on
the object. Diagnostics carry only a program name, io::ErrorKind and
errno, matching ProgramLoadRejected's existing redaction contract.

The READMEs contradicted themselves in one sentence: a bpf_loop
availability note scoped to mainline 5.17 read as a floor, while the
next clause asserted a 6.8 profile, and neither governed a kernel that
has the helper but predates 6.8. Both files now say helper availability
is not loadability, that CI proves exactly one kernel, and that other
kernels are unqualified rather than known-good or known-bad, pointing at
the probe instead of a version comparison.

The 6.8 gate could also pass without loading anything: both privileged
tests return early when OPC_GTPU_RUN_PRIVILEGED is unset, and a skipped
run still prints "1 passed". Each now prints a marker only after a real
load and CI asserts both, and the gate covers the new probe so the
no-residue claim is proven on a real kernel.

Refs #543

* ci(gtpu-dataplane): gate the classifier load on the OpenShift 4.18 kernel

OpenShift 4.18 runs RHCOS on RHEL 9.4, so its nodes are the 5.14.0 line
rather than the 6.8 mainline kernel CI proved. That line is an
enterprise backport: it carries bpf_loop, but helper presence does not
establish that this object's callback chain fits the verifier's
cumulative 512-byte stack accounting, which is the only property that
decides whether the datapath loads. Stating the line as unqualified was
therefore the wrong answer for a supported deployment target.

Gate it instead. A second QEMU job boots a Rocky 9 guest -- ABI
equivalent to RHEL 9 and publicly downloadable, unlike RHEL itself --
asserts the guest really is on 5.14 before trusting anything, and loads
both committed classifiers plus the capability probe there.

The blocker was binary portability, not the kernel: the test binary is
built on an Ubuntu runner and executed unmodified in the guest, and a
gnu build will not start against el9's glibc. Building for
x86_64-unknown-linux-musl produces a static binary that runs on either,
and the job asserts it is statically linked so a silent regression to a
dynamic build cannot pass.

That build needed one portability fix: statfs::f_type is signed on musl
and unsigned on glibc, so the bpffs magic comparison now widens both
sides instead of assuming one libc.

Both READMEs now record 5.14/el9 as a gated line rather than an
unqualified one, and keep the distinction the issue asked for: helper
availability is not loadability, and kernels outside the gated lines
still have no verdict.

Refs #543

* ci(gtpu-dataplane): assert el9 binary staticness by PT_INTERP

The el9 gate asserted 'statically linked' against file(1), which reports
'static-pie linked' for a static musl PIE; only ldd(1) uses the other
phrasing. The build and the binary were correct and the job still failed
before reaching the load. Check for an absent program interpreter
instead, which is the property that decides whether the binary starts on
RHEL 9 and does not depend on either tool's wording.

* fix(gtpu-dataplane): prove the OpenShift 4.18 kernel, not a newer one

Adversarial review found the gate did not prove what its name claimed,
and that the capability API would mislead an operator on exactly the
platform it targets.

The gate asserted only that the guest was on 5.14 and pulled Rocky's
floating latest image, so the green run proved loadability on
5.14.0-687.el9_8 -- roughly four RHEL minor rebases past the
5.14.0-427 el9_4 kernel RHCOS ships for OpenShift 4.18. The gate's own
premise is that an enterprise backport can account for the bpf_loop
callback chain differently, and that argument applies between el9
z-streams too, so the job refuted its own conclusion. It now boots a
digest-pinned Rocky 9.4 image and asserts 5.14.0-427.*el9_4*, plus the
vendor and virtualisation checks the 6.8 job already made. The image was
also the only download in the repository without a checksum, which made
the load-bearing evidence neither reproducible nor integrity-checked.

aya funnels every BPF_PROG_LOAD failure into one error variant without
inspecting errno, so the probe reported an LSM denial as
VerifierRejected -- telling an operator the node can never forward when
the fix is an SELinux policy change. That is the normal state on RHCOS:
a pod can hold CAP_SYS_ADMIN, satisfy the capability precheck, and still
be denied bpf { prog_load }. The kernel only writes a verifier log when
it ran the verifier, so an empty log now routes PermissionDenied to
InsufficientPrivileges and anything else to Environment. The log is
still dropped rather than retained.

The READMEs claimed the line was supported when two load tests are not
attach and not forwarding, and left a production node profile remains
Linux 6.8 or newer three lines above the new bullet, which would send a
reader straight back to the version comparison this work exists to
replace. They now state what the gate proves, what it does not, and that
a loadability verdict does not override the kernel floor
opc-node-resources applies to gateway profiles.

Also: the unattemptable-environment test probed /nonexistent, and since
the probe creates missing parents it littered the filesystem root when
run as root; it uses a temp dir and asserts the classification instead.
The Environment blocker now documents that a CO-RE relocation failure
lands there and is not necessarily fixable, and the libc signedness
comment no longer states the pairing backwards.

Refs #543

* docs(gtpu-dataplane): scope the loadability verdict correctly

The previous wording implied opc-node-resources applies a kernel floor a
node must satisfy alongside this verdict. It does not: the floor lives
in IpsecGatewayProfile::standard(), which is a constructor used only by
tests and the hardware-lab testbed, has no Default impl, and gates the
IPsec path rather than the GTP-U datapath. The validator compares
against whatever minimum the caller supplies, so nothing in the SDK
refuses a 5.14 node by default. Say what is true instead: this verdict
covers the object, not node admission.

+++ crates/opc-linux-xfrm-sys/src/linux.rs
@@ -121,7 +121,10 @@ fn verify_bpffs(fd: &OwnedFd) -> io::Result<()> {
     if unsafe { libc::fstatfs(fd.as_raw_fd(), &mut status) } < 0 {
         return Err(io::Error::last_os_error());
     }
-    if status.f_type != BPF_FS_MAGIC {
+    // The signedness of `statfs::f_type` and of `BPF_FS_MAGIC` differs by
+    // libc and target, so widen both sides into a type that holds either
+    // rather than assuming a particular pairing.
+    if i128::from(status.f_type) != i128::from(BPF_FS_MAGIC) {
         return Err(io::Error::new(
             io::ErrorKind::InvalidInput,
             "pin directory is not on bpffs",
